fix: ingame menu not saving position correctly
For example, when menu is closed while guide is open, we end up
selecting the wrong entry when the menu is opened again. This is because
we choose the index of the menu based on the enum ordinal, but the menu
is dynamic. This fixes so we choose the index based on the actual menu
entries.